refactor: rename `position` module to `span`
