+++ frontend/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod lexer;
-mod position;
 mod scanner;
+mod span;
 mod symbol;
